# projectT4T/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/Regist": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Regist User",
                "operationId": "UsersRegist",
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Regist User",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "nama"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "nama@mail.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "nama123"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/Login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login User",
                "operationId": "UsersLogin",
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Login User",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "nama@mail.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "nama123"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/Logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Logout User",
                "operationId": "UsersLogout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/ForgotPassword": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Forgot Password User",
                "operationId": "ForgotPassword",
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Forgot Password User",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "nama@mail.com"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/EditProfile": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Edit Profile User",
                "operationId": "EditProfile",
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Edit Profile User",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "nama"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "nama@mail.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "1QwOp@"
                                },
                                "new_password": {
                                    "type": "string",
                                    "example": "1QwOp@"
                                },
                                "confirm_password": {
                                    "type": "string",
                                    "example": "1QwOp@"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
...
    "securityDefinitions": {
        "apiAuth": {
            "type": "apiKey",
            "description": "Login with email and password to get the authentication token",
            "name": "Authorization",
            "in": "header"
        }
    }
}